+++ drasyl-plugin-groups-manager/src/test/java/org/drasyl/plugin/groups/manager/data/GroupTest.java
@@ -36,28 +36,28 @@ class GroupTest {
     class Getters {
         @Test
         void shouldReturnCorrectName() {
-            final Group group = Group.of("test", "secret", (short) 0, ofSeconds(60));
+            final Group group = Group.of("test", "secret", (byte) 0, ofSeconds(60));
 
             assertEquals("test", group.getName());
         }
 
         @Test
         void shouldReturnCorrectSecret() {
-            final Group group = Group.of("test", "secret", (short) 0, ofSeconds(60));
+            final Group group = Group.of("test", "secret", (byte) 0, ofSeconds(60));
 
             assertEquals("secret", group.getCredentials());
         }
 
         @Test
         void shouldReturnCorrectMinDifficulty() {
-            final Group group = Group.of("test", "secret", (short) 0, ofSeconds(60));
+            final Group group = Group.of("test", "secret", (byte) 0, ofSeconds(60));
 
             assertEquals(0, group.getMinDifficulty());
         }
 
         @Test
         void shouldReturnCorrectTimeout() {
-            final Group group = Group.of("test", "secret", (short) 0, ofSeconds(60));
+            final Group group = Group.of("test", "secret", (byte) 0, ofSeconds(60));
 
             assertEquals(ofSeconds(60), group.getTimeout());
         }
@@ -67,16 +67,16 @@ class GroupTest {
     class Equals {
         @Test
         void shouldBeEquals() {
-            final Group group1 = Group.of("test", "secret", (short) 0, ofSeconds(60));
-            final Group group2 = Group.of("test", "secret", (short) 0, ofSeconds(60));
+            final Group group1 = Group.of("test", "secret", (byte) 0, ofSeconds(60));
+            final Group group2 = Group.of("test", "secret", (byte) 0, ofSeconds(60));
 
             assertEquals(group1, group2);
         }
 
         @Test
         void shouldNotBeEquals() {
-            final Group group1 = Group.of("test1", "secret", (short) 0, ofSeconds(60));
-            final Group group2 = Group.of("test2", "secret", (short) 0, ofSeconds(60));
+            final Group group1 = Group.of("test1", "secret", (byte) 0, ofSeconds(60));
+            final Group group2 = Group.of("test2", "secret", (byte) 0, ofSeconds(60));
 
             assertNotEquals(group1, group2);
         }
@@ -86,16 +86,16 @@ class GroupTest {
     class HashCode {
         @Test
         void shouldBeEquals() {
-            final Group group1 = Group.of("test", "secret", (short) 0, ofSeconds(60));
-            final Group group2 = Group.of("test", "secret", (short) 0, ofSeconds(60));
+            final Group group1 = Group.of("test", "secret", (byte) 0, ofSeconds(60));
+            final Group group2 = Group.of("test", "secret", (byte) 0, ofSeconds(60));
 
             assertEquals(group1.hashCode(), group2.hashCode());
         }
 
         @Test
         void shouldNotBeEquals() {
-            final Group group1 = Group.of("test1", "secret", (short) 0, ofSeconds(60));
-            final Group group2 = Group.of("test2", "secret", (short) 0, ofSeconds(60));
+            final Group group1 = Group.of("test1", "secret", (byte) 0, ofSeconds(60));
+            final Group group2 = Group.of("test2", "secret", (byte) 0, ofSeconds(60));
 
             assertNotEquals(group1.hashCode(), group2.hashCode());
         }
@@ -106,13 +106,13 @@ class GroupTest {
         @Test
         void shouldThrowExceptionOnInvalidMinDifficulty() {
             final Duration timeout = ofSeconds(60);
-            assertThrows(IllegalArgumentException.class, () -> Group.of("vip-gang", "secret", (short) -1, timeout));
+            assertThrows(IllegalArgumentException.class, () -> Group.of("vip-gang", "secret", (byte) -1, timeout));
         }
 
         @Test
         void shouldThrowExceptionOnInvalidTimeout() {
             final Duration timeout = ofSeconds(1);
-            assertThrows(IllegalArgumentException.class, () -> Group.of("vip-gang", "secret", (short) 0, timeout));
+            assertThrows(IllegalArgumentException.class, () -> Group.of("vip-gang", "secret", (byte) 0, timeout));
         }
     }
 }
